Start using the new SR1 S2 correction map (v2.2) (#154)
* Start using the new SR1 S2 correction map (v2.2)

This causes the corrections treemaker to start using the v2.2 s2 map which is added to pax in pull request pax/#626. The map is used for science run 1 data only (run number greater than 6386). For science run 0 the map is unchanged.

It should not be merged until the new map has been added to pax.

* Increment Corrections treemaker version

Increment the corrections treemaker version so that minitrees are regenerated