Bump Corrections 2.3 --> 2.4 (#260)
This bumps the Corrections treemaker from 2.3 to 2.4 due to updates to the electron lifetime model and new TFNN map for ~3 months of SR2. We're still waiting on the alpha electron lifetime study, but I think we should go ahead and grow minitrees so that these updates can be used in the meantime. Does anyone oppose this?